fix: goal 생성 테스트 class 변경
